% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{calculate_visual_angle}
\alias{calculate_visual_angle}
\title{Calculate visual angle from physical measurements}
\usage{
calculate_visual_angle(size, distance)
}
\arguments{
\item{size}{Physical size (width or height) in any unit}

\item{distance}{Viewing distance in the SAME unit as size}
}
\value{
Visual angle in degrees
}
\description{
Converts physical screen dimensions and viewing distance to visual angle
in degrees. Uses the formula: angle = 2 * atan(size / (2 * distance)) * (180/pi)
}
