% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_gaze_summary_plot}
\alias{create_gaze_summary_plot}
\title{Create summary visualization of gaze patterns}
\usage{
create_gaze_summary_plot(nifti_data, integrated_data)
}
\arguments{
\item{nifti_data}{List from preload_nifti_data()}

\item{integrated_data}{Data frame from integrate_all_gaze_points()}
}
\description{
Create summary visualization of gaze patterns
}
