% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti-io.R
\name{fast_slice_viewer}
\alias{fast_slice_viewer}
\title{Fast slice viewer using pre-extracted data}
\usage{
fast_slice_viewer(data, slice = NULL)
}
\arguments{
\item{data}{List from preload_nifti_data()}

\item{slice}{Slice number to display}
}
\description{
Fast slice viewer using pre-extracted data
}
