% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate-mapping.R
\name{plot_coordinate_mapping}
\alias{plot_coordinate_mapping}
\title{Create coordinate mapping summary plot}
\usage{
plot_coordinate_mapping(locations, nifti_data)
}
\arguments{
\item{locations}{Data frame from map_gaze_to_anatomy()}

\item{nifti_data}{List from preload_nifti_data()}
}
\description{
Create coordinate mapping summary plot
}
