% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti-io.R
\name{preload_nifti_data}
\alias{preload_nifti_data}
\title{Load NIfTI file with pre-extracted data}
\usage{
preload_nifti_data(filepath)
}
\arguments{
\item{filepath}{Path to the NIfTI file}
}
\value{
A list containing:
\item{nifti}{The RNifti object}
\item{nvimage}{An NVImage object with transformation matrices}
\item{data}{Pre-extracted image data as array}
\item{dims}{Dimensions of the image}
}
\description{
Loads a NIfTI file and pre-extracts all image data for faster access.
Also creates an NVImage object for coordinate transformations.
}
