% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti-io.R
\name{safe_coordinate_picker}
\alias{safe_coordinate_picker}
\title{Interactive coordinate picker for NIfTI slices}
\usage{
safe_coordinate_picker(data, slice = NULL)
}
\arguments{
\item{data}{List from preload_nifti_data()}

\item{slice}{Slice number to display}
}
\value{
Data frame of clicked points with coordinates
}
\description{
Allows interactive clicking on a NIfTI slice to get voxel and world coordinates.
Right-click to exit.
}
