% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti-io.R
\name{safe_get_slice}
\alias{safe_get_slice}
\title{Safely extract a slice from NIfTI data}
\usage{
safe_get_slice(nii, slice_num)
}
\arguments{
\item{nii}{NIfTI object}

\item{slice_num}{Slice number to extract}
}
\value{
Matrix of slice data or NULL if error
}
\description{
Safely extract a slice from NIfTI data
}
