% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate-mapping.R
\name{safe_get_value}
\alias{safe_get_value}
\title{Safely extract intensity value from image data}
\usage{
safe_get_value(img_data, voxel, dims)
}
\arguments{
\item{img_data}{3D array of image data}

\item{voxel}{Voxel coordinates (1-based)}

\item{dims}{Image dimensions}
}
\value{
Intensity value or NA if out of bounds
}
\description{
Safely extract intensity value from image data
}
