% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifti-io.R
\name{view_multiple_slices}
\alias{view_multiple_slices}
\title{View multiple NIfTI slices in a grid}
\usage{
view_multiple_slices(data, slices = NULL)
}
\arguments{
\item{data}{List from preload_nifti_data()}

\item{slices}{Vector of slice numbers (NULL for equally spaced)}
}
\description{
View multiple NIfTI slices in a grid
}
